src/requirements.txt
anyio==3.6.2
arrow==1.2.3
binaryornot==0.4.4
chardet==3.0.2
cookiecutter==2.1.1
email-validator==1.3.0
fastapi==0.86.0
h11==0.8.0
http3==0.6.7
httptools==0.5.0
jinja2-time==0.2.0
names==0.3.0
pydantic==1.10.2
psycopg2==2.9.5
cryptography==38.0.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
sniffio==1.3.0
starlette==0.20.4
SQLAlchemy-Utils==0.38.3
text-unidecode==1.3
typing-extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
